refactor(tauri): move media helpers to utils
