+++ apps/desktop/src-tauri/src/utils/media.rs
