+++ apps/desktop/src-tauri/src/utils/mod.rs
@@ -2,3 +2,5 @@ pub mod date;
 pub mod file_ops;
 pub mod file_utils;
 pub mod identifier;
+pub mod media;
+pub mod screen_capture;
